Remove non-hardened items from TON derivation path Add android (kotlin) tests Rename expiredAt => expireAt and make it uint64 in proto Other minor code style fixes

+++ src/TheOpenNetwork/wallet/Wallet.h
@@ -39,13 +39,13 @@ public:
         uint64_t amount,
         uint32_t seqno,
         uint8_t mode,
-        uint32_t expireAt = 0,
+        uint64_t expireAt = 0,
         const std::string& comment = ""
     ) const;
 
 protected:
     [[nodiscard]] virtual Cell::Ref createDataCell() const = 0;
-    virtual void writeSigningPayload(CellBuilder& builder, uint32_t seqno = 0, uint32_t expireAt = 0) const = 0;
+    virtual void writeSigningPayload(CellBuilder& builder, uint32_t seqno = 0, uint64_t expireAt = 0) const = 0;
 
 private:
     [[nodiscard]] Cell::Ref createSigningMessage(
@@ -53,7 +53,7 @@ private:
         uint64_t amount,
         uint32_t seqno,
         uint8_t mode,
-        uint32_t expireAt = 0,
+        uint64_t expireAt = 0,
         const std::string& comment = ""
     ) const;
     [[nodiscard]] Everscale::StateInit createStateInit() const;
